+++ src/test/kotlin/io/libp2p/simulate/util/MessageSizeEstimator.kt
@@ -2,11 +2,11 @@ package io.libp2p.simulate.util
 
 import io.netty.buffer.ByteBuf
 
-typealias MsgSizeEstimator = (Any) -> Int
+typealias MsgSizeEstimator = (Any) -> Long
 
 val GeneralSizeEstimator: MsgSizeEstimator = { msg ->
     when (msg) {
-        is ByteBuf -> msg.readableBytes()
+        is ByteBuf -> msg.readableBytes().toLong()
         else -> 0
     }
 }
